system: characterization tests, and the SQLite reads move to their own repository
